Put getControllerNames in VehicleControllerType
