+++ src/test/java/ch/heigvd/sitr/model/SimulationTest.java
@@ -12,19 +12,8 @@ import java.util.LinkedList;
 
 /**
  * Unit tests for Simulation class
+ *
  * @author Luc Wachter
  */
 class SimulationTest {
-    /**
-     * Does the getControllerNames method return a correct List of available controllers?
-     */
-    @Test
-    public void shouldEnableToGetControllerNames() {
-        LinkedList<String> controllerNames = new LinkedList<>();
-        for (VehicleControllerType vct : VehicleControllerType.values()) {
-            controllerNames.add(vct.getName());
-        }
-
-        assertEquals(controllerNames, Simulation.getControllerNames());
-    }
 }
