+++ src/test/java/ch/heigvd/sitr/model/VehicleControllerTypeTest.java
@@ -6,10 +6,14 @@
 package ch.heigvd.sitr.model;
 
 import org.junit.jupiter.api.Test;
+
+import java.util.LinkedList;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Unit tests for the VehicleControllerType enum
+ *
  * @author Luc Wachter
  */
 class VehicleControllerTypeTest {
@@ -26,5 +30,20 @@ class VehicleControllerTypeTest {
     public void configPathGetterShouldWork() {
         assertEquals("config/vehicleController/careful.xml", VehicleControllerType.CAREFUL.getConfigPath());
         assertEquals("config/vehicleController/reckless.xml", VehicleControllerType.RECKLESS.getConfigPath());
+    }
+
+    /**
+     * Does the getControllerNames method return a correct List of available controllers?
+     */
+    @Test
+    public void shouldEnableToGetControllerNames() {
+        // Test manually
+        LinkedList<String> controllerNames = new LinkedList<>();
+        controllerNames.add("Timid");
+        controllerNames.add("Careful");
+        controllerNames.add("Reckless");
+        controllerNames.add("Autonomous");
+
+        assertEquals(controllerNames, VehicleControllerType.getControllerNames());
     }
 }
